+++ HISTORY.rst
@@ -174,3 +174,4 @@ History
 
 * Fix for when the period passed to manual tip filter contains a nan
 * Another edge case fix for Null manual tips (caused floats)
+* Fixed quality series adding values from before start date